[히로]로또 2단계 - 수동 미션 제출합니다. (#220)
* refactor: LottoMachine 싱글톤 패턴

* refactor: 상수 중복 제거

* refactor: String 리터럴을 상수로

* feat: lottoNumber로 포장

* feat: LottoMachine에서 한번만 생성한 LottoNumber를 계속 사용

* refactor: Lotto가 원시값 대신 LottoNumber를 사용하도록 수정

* refactor: LottoDTO, LottoDTOs의 toString() 오버라이딩

* refactor: WinningLotto의 중복된 검사 로직 제거

* refactor: WinningLotto의 불필요한 메소드 분리 제거

* refactor: 줄바꿈 및 리터럴 상수화 수정

* refactor: 수익률 소수점 자리까지 표현하게 수정

* feat: LottoMachine에서 수동 로또 뽑기 구현

* feat: ~WinningNumber -> ~LottoNumber

자동 로또에서는 당첨 번호만 수동입력 받았으나, 수동 로또 구현을 위해 이름 변경

* refactor: WinningLotto의 검증 부담을 Lotto와 LottoNumber가 지도록 통합

* docs: README.md 요구사항 업데이트

* feat: LottoCountException, LottosException, WinningLottoException 추가

* feat: LottoCountException, LottosException, WinningLottoException 추가

* feat: LottoCount에서 수동 로또 구매 장수를 처리

* feat: LottoController에서 수동 로또를 구매할 수 있도록 구현

* refactor: 수동 로또 번호 입력 도중 에러가 나도 처음으로 튕기지 않게 수정

* fix: 두 개 이상의 로또가 일등 당첨될 경우 상금이 Integer 범위를 넘어가는 문제 수정

* refactor: Lotto.refactor() 구현 단순화

* refactor: WinningLottoTest 구현사항이 없으므로 삭제

* refactor: buyManualLottos 분리

* docs: README.md 수정 사항 반영

* fix: LottoMachineTest 테스트 통과 못하는 문제 수정

* refactor: 상수 구현 수정, 중복 제거

* refactor: lotto.toString() -> lotto

* refactor: lottoDtos -> lottosDto

* refactor: 중복 구현한 LottoMachine.pickDedicatedBalls()를 활용

* refactor: Controller의 입력값 처리 기능을 InputUtil이 부담하도록 수정

* refactor: LottoDtos -> LottosDto 일부 바뀌지 않은 부분 수정.

* refactor: 목적에 충실하게 간결하게. 불필요한 상수화 수정

* refactor: LottoCountTest에 정상적으로 로또가 발급될 경우에 대한 테스트 추가

* refactor: Lotto, LottoNumber: 메시지 상수화 시키지 않도록 변경.

* refactor: LottoMachine에서 exception 안내 메시지 수정.

* refactor: InputUtil -> InputView

* refactor: LottoDto의 toString() 구현 간소화

+++ src/main/java/lotto/domain/WinningPrize.java
@@ -11,7 +11,7 @@ public enum WinningPrize {
 	FIRST(6, 2_000_000_000, "1등");
 
 	private int matchCount;
-	private int prize;
+	private long prize;
 	private String description;
 
 	WinningPrize(int matchCount, int prize, String description) {
@@ -25,12 +25,12 @@ public enum WinningPrize {
 			return SECOND;
 		}
 		return Arrays.stream(values())
-			.filter(prize -> prize.matchCount == matchCount)
-			.findFirst()
-			.orElse(NO_PRIZE);
+				.filter(prize -> prize.matchCount == matchCount)
+				.findFirst()
+				.orElse(NO_PRIZE);
 	}
 
-	public int getPrize() {
+	public long getPrize() {
 		return prize;
 	}
